+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # LAST CHANGES:
 
+Version `v1.7.1 (29 Feb'24)`
+
+  - BUGFIX: Moving unit descriptions: processed tabs in tabbar not marked as modified
+  - CHANGE: Moving unit descriptions: "All in session" instead of "All open"
+  - CHANGE: Moving unit descriptions: intro lines ("All of...") skipped
+
+
 Version `v1.7.0 (28 Feb'24)`
 
   - BUGFIX: run e_menu tools on file.c